+++ RandomSkunk.Results/Error.cs
@@ -4,47 +4,28 @@ namespace RandomSkunk.Results;
 /// Defines an error.
 /// </summary>
 [DebuggerDisplay($"{{{nameof(GetDebuggerDisplay)}(),nq}}")]
-public class Error : IEquatable<Error>
+public record class Error
 {
-    private static readonly Lazy<Error> _default = new(() => new Error());
+    private static readonly Lazy<Error> _defaultError = new(() => new Error());
 
     private static string _defaultMessage = "An error occurred.";
+
+    private readonly string _message;
+    private readonly string _type;
 
     /// <summary>
     /// Initializes a new instance of the <see cref="Error"/> class.
     /// </summary>
     /// <param name="message">
-    /// The optional error message. If <see langword="null"/>, then the value of
-    /// <see cref="DefaultMessage"/> is used instead.
+    /// The error message. If <see langword="null"/>, then the value of <see cref="DefaultMessage"/> is used instead.
     /// </param>
-    /// <param name="stackTrace">The optional stack trace.</param>
-    /// <param name="errorCode">The optional error code.</param>
-    /// <param name="identifier">The optional identifier of the error.</param>
     /// <param name="type">
-    /// The optional type of the error. If <see langword="null"/>, then the
-    /// <see cref="MemberInfo.Name"/> of the <see cref="System.Type"/> of the current instance
-    /// is used instead.
+    /// The type of the error. If <see langword="null"/>, then the name of the error type is used instead.
     /// </param>
-    /// <param name="innerError">
-    /// The optional error that is the cause of the current error.
-    /// </param>
-    /// <exception cref="ArgumentNullException">
-    /// If <paramref name="message"/> is <see langword="null"/>.
-    /// </exception>
-    public Error(
-        string? message = null,
-        string? stackTrace = null,
-        int? errorCode = null,
-        string? identifier = null,
-        string? type = null,
-        Error? innerError = null)
+    public Error(string? message = null, string? type = null)
     {
-        Message = message ?? DefaultMessage;
-        StackTrace = stackTrace;
-        ErrorCode = errorCode;
-        Identifier = identifier;
-        Type = type ?? GetType().Name;
-        InnerError = innerError;
+        _message = message ?? _defaultMessage;
+        _type = type ?? GetType().Name;
     }
 
     /// <summary>
@@ -62,59 +43,43 @@ public class Error : IEquatable<Error>
     /// <summary>
     /// Gets the error message.
     /// </summary>
-    public string Message { get; }
+    public string Message
+    {
+        get => _message;
+        init => _message = value ?? throw new ArgumentNullException(nameof(value));
+    }
 
     /// <summary>
     /// Gets the optional stack trace.
     /// </summary>
-    public string? StackTrace { get; }
+    public string? StackTrace { get; init; }
 
     /// <summary>
     /// Gets the optional error code.
     /// </summary>
-    public int? ErrorCode { get; }
+    public int? ErrorCode { get; init; }
 
     /// <summary>
     /// Gets the optional identifier of the error.
     /// </summary>
-    public string? Identifier { get; }
+    public string? Identifier { get; init; }
 
     /// <summary>
     /// Gets the type of the error.
     /// </summary>
-    public string Type { get; }
+    public string Type
+    {
+        get => _type;
+        init => _type = value ?? throw new ArgumentNullException(nameof(value));
+    }
 
     /// <summary>
     /// Gets the optional <see cref="Error"/> instance that caused the current error.
     /// </summary>
     [JsonConverter(typeof(InnerErrorJsonConverter))]
-    public Error? InnerError { get; }
+    public Error? InnerError { get; init; }
 
-    internal static Error DefaultError => _default.Value;
-
-    /// <summary>
-    /// Indicates whether the <paramref name="left"/> parameter is equal to the
-    /// <paramref name="right"/> parameter.
-    /// </summary>
-    /// <param name="left">The left side of the comparison.</param>
-    /// <param name="right">The right side of the comparison.</param>
-    /// <returns>
-    /// <see langword="true"/> if the <paramref name="left"/> parameter is equal to the
-    /// <paramref name="right"/> parameter; otherwise, <see langword="false"/>.
-    /// </returns>
-    public static bool operator ==(Error? left, Error? right) => ReferenceEquals(left, right) || (left is not null && left.Equals(right));
-
-    /// <summary>
-    /// Indicates whether the <paramref name="left"/> parameter is not equal to the
-    /// <paramref name="right"/> parameter.
-    /// </summary>
-    /// <param name="left">The left side of the comparison.</param>
-    /// <param name="right">The right side of the comparison.</param>
-    /// <returns>
-    /// <see langword="true"/> if the <paramref name="left"/> parameter is not equal to the
-    /// <paramref name="right"/> parameter; otherwise, <see langword="false"/>.
-    /// </returns>
-    public static bool operator !=(Error? left, Error? right) => !(left == right);
+    internal static Error DefaultError => _defaultError.Value;
 
     /// <summary>
     /// Creates an <see cref="Error"/> object from the specified <see cref="Exception"/>.
@@ -152,38 +117,17 @@ public class Error : IEquatable<Error>
         else
             message += Environment.NewLine + exceptionMessage;
 
-        return new Error(message, exception.StackTrace, errorCode, identifier, type, innerError);
+        return new Error(message, type)
+        {
+            StackTrace = exception.StackTrace,
+            ErrorCode = errorCode,
+            Identifier = identifier,
+            InnerError = innerError,
+        };
     }
 
     /// <inheritdoc/>
     public override string ToString() => ToString(string.Empty);
-
-    /// <inheritdoc/>
-    public virtual bool Equals(Error? other) =>
-        other is not null
-            && Message == other.Message
-            && StackTrace == other.StackTrace
-            && ErrorCode == other.ErrorCode
-            && Identifier == other.Identifier
-            && Type == other.Type
-            && InnerError == other.InnerError;
-
-    /// <inheritdoc/>
-    public override bool Equals(object? obj) =>
-        obj is Error other && Equals(other);
-
-    /// <inheritdoc/>
-    public override int GetHashCode()
-    {
-        int hashCode = 910147539;
-        hashCode = (hashCode * -1521134295) + Message.GetHashCode();
-        hashCode = (hashCode * -1521134295) + (StackTrace is null ? 0 : StackTrace.GetHashCode());
-        hashCode = (hashCode * -1521134295) + ErrorCode.GetHashCode();
-        hashCode = (hashCode * -1521134295) + (Identifier is null ? 0 : Identifier.GetHashCode());
-        hashCode = (hashCode * -1521134295) + Type.GetHashCode();
-        hashCode = (hashCode * -1521134295) + (InnerError is null ? 0 : InnerError.GetHashCode());
-        return hashCode;
-    }
 
     /// <summary>
     /// When overridden in an inherited class, appends additional properties of the derived error class to the