+++ RandomSkunk.Results/Polyfill/DisallowNullAttribute.cs
@@ -3,6 +3,7 @@
 namespace System.Diagnostics.CodeAnalysis;
 
 [AttributeUsage(AttributeTargets.Property | AttributeTargets.Field | AttributeTargets.Parameter, Inherited = false)]
+[EditorBrowsable(EditorBrowsableState.Never)]
 internal sealed class DisallowNullAttribute : Attribute
 {
 }