+++ RandomSkunk.Results/Polyfill/IsExternalInit.cs
@@ -1,0 +1,10 @@
+#if NET461 || NETSTANDARD2_0 || NETSTANDARD2_1
+
+namespace System.Runtime.CompilerServices;
+
+[EditorBrowsable(EditorBrowsableState.Never)]
+internal static class IsExternalInit
+{
+}
+
+#endif