+++ RandomSkunk.Results/Polyfill/NotNullWhenAttribute.cs
@@ -3,6 +3,7 @@
 namespace System.Diagnostics.CodeAnalysis;
 
 [AttributeUsage(AttributeTargets.Parameter, Inherited = false)]
+[EditorBrowsable(EditorBrowsableState.Never)]
 internal sealed class NotNullWhenAttribute : Attribute
 {
     public NotNullWhenAttribute(bool returnValue) => ReturnValue = returnValue;